Create Player State Management Project
Implement Input Handler class for handling
different inputs.

Implement Player class with different types of
movements and animations controlled by a set of
states in a state machine mechanishm.

Implement State class which stores all data and
functionality of each state the player can be in.

Implement Status Text functionality on canvas to display
the key pressed and the current state the player is in.

Create animation loop for canvas that will repeatedly
erase and draw its content at each frame and create the
illusion of animation.

Animate player movement and state based off the layout of
the sprite sheet provided.